refactor(types): Cambiar double a float en estructuras de coordenadas
- Vertex: double lat/lon -> float (16B -> 8B, 50% reduccion)
- Position: double lat/lon -> float (16B -> 8B)
- XY: double x/y -> float
- Cow::distanceToLimit: double -> float
- MAX_VERTICES: 20 -> 10 (reduccion 50%)

Justificacion: GPS u-blox SAM-M10Q tiene precision 1.5m CEP.
Float (precision ~1.1m) es mas que suficiente vs double (~1mm).
Ahorro: 1068 bytes RAM total.

+++ CM4/Core/Inc/Types/cow.h
@@ -18,8 +18,8 @@ enum class CowState {
 };
 
 struct Position {
-    double latitude;
-    double longitude;
+    float latitude;
+    float longitude;
 };
 
 struct Acceleration {
@@ -36,14 +36,14 @@ public:
     void updateAcceleration(Acceleration accel);
     void updateState(CowState state);
     void updateCurrentZone(zone_t zone);
-    void updateDistanceToLimit(double distance);
+    void updateDistanceToLimit(float distance);
 
     DeviceUID getId() const;
     Position getPosition() const;
     Acceleration getAcceleration() const;
     CowState getState() const;
     zone_t getCurrentZone() const;
-    double getDistanceToLimit() const;
+    float getDistanceToLimit() const;
 
 private:
     DeviceUID id;
@@ -51,7 +51,7 @@ private:
     Acceleration acceleration;
     CowState state;
     zone_t currentZone;
-    double distanceToLimit;
+    float distanceToLimit;
 };
 
 #endif
